clearing out johnny-five and adding boilerplate for aruidno code
