+++ package.json
@@ -5,7 +5,6 @@
   "main": "index.js",
   "dependencies": {
     "dotenv": "^8.2.0",
-    "johnny-five": "^2.0.0",
     "node-fetch": "^2.6.1",
     "serialport": "^9.0.7"
   },
