+++ test.js
@@ -1,18 +1,13 @@
-// manually update a pin
-// used for testing
+// manually update pins to test endpoint
 
 require("dotenv").config();
 const fetch = require("node-fetch");
 
 function getRandomNumber() {
   const min = Math.ceil(0);
-  const max = Math.floor(1023);
+  const max = Math.floor(255);
   return Math.floor(Math.random() * (max - min) + min);
 }
-// const pin = {
-//   id: "A0",
-//   value: 1200,
-// };
 
 const allPins = [
   {
@@ -24,7 +19,7 @@ const allPins = [
     value: getRandomNumber(),
   },
   {
-    id: "A2",
+    id: "D2",
     value: getRandomNumber(),
   },
 ];
@@ -40,22 +35,3 @@ for (const pin of allPins) {
     .then((json) => console.log(json))
     .catch((err) => console.log(err));
 }
-// updates pin A0
-// fetch(`${process.env.API_HOST}/pins/A0`, {
-//   method: "PUT",
-//   body: JSON.stringify(pin),
-//   headers: { "Content-Type": "application/json" },
-// })
-//   .then((res) => res.json())
-//   .then((json) => console.log(json))
-//   .catch((err) => console.log(err));
-
-// // update multiple pins
-// fetch(`${process.env.API_HOST}/pins/A1`, {
-//   method: "PUT",
-//   body: JSON.stringify(allPins[1]),
-//   headers: { "Content-Type": "application/json" },
-// })
-//   .then((res) => res.json())
-//   .then((json) => console.log(json))
-//   .catch((err) => console.log(err));
